+++ NEWS.md
@@ -1,3 +1,8 @@
+# graphTweets 0.5.2
+
+* `gt_add_meta` Added to add custom meta data to nodes. 
+* `gt_preproc_edges` Added to preprocess edges before comoputing nodes.
+
 # graphTweets 0.5.1
 
 Better use of tidyeval to improve usage of the package in functions.